Documentation + Moved some functions around
