+++ pymuonsuite/quantum/grid.py
@@ -92,40 +92,3 @@ def weighted_tens_avg(tensors, weight):
         tens_avg[i] = np.sum(
             weight[:, :, None, None]*tensors[i], axis=(0, 1))/np.sum(weight)
     return tens_avg
-
-def write_tensors(tensors, sname, symbols):
-    """
-    Write out a set of 3x3 tensors for every atom in a system.
-
-    | Args:
-    |   tensors(Numpy float array, shape: (Atoms, 3, 3): A list of 3x3 tensors
-    |       for each atom.
-    |   sname(str): Seedname for file (i.e. filename will be sname_tensors.dat).
-    |   symbols(str array): List containing chemical symbol of each atom in
-    |       system.
-    |
-    | Returns: Nothing
-    """
-    tensfile = open(sname + '_tensors.dat', 'w')
-    for i in range(np.size(tensors, 0)):
-        tensfile.write('{0} {1}\n'.format(symbols[i], i))
-        tensfile.write('\n'.join(['\t'.join([str(x) for x in l]) for l in tensors[i]]) + '\n')
-
-def calc_harm_potential(R, grid_n, mu_mass, freqs, E_table, sname):
-    """
-    Calculate the harmonic potential at all displacements on the grid and write
-    out to file in a format that can be plotted.
-    """
-    R_axes = np.array([np.linspace(-3*Ri, 3*Ri, grid_n)
-                       for Ri in R])
-    # Now the potential, measured vs. theoretical
-    harm_K = mu_mass*freqs**2
-    harm_V = (0.5*harm_K[:, None]*(R_axes*1e-10)**2)/cnst.electron_volt
-    # Normalise E_table
-    if E_table.shape[1] % 2 == 1:
-        E_table -= (E_table[:, E_table.shape[1]//2])[:, None]
-    else:
-        E_table -= (E_table[:, E_table.shape[1]//2] +
-                    E_table[:, E_table.shape[1]//2-1])[:, None]/2.0
-    all_table = np.concatenate((R_axes, harm_V, E_table), axis=0)
-    np.savetxt(sname + '_V.dat', all_table.T)
